requirements.txt
django == 5.1.2

djangorestframework==3.15.2


#scraping
beautifulsoup4==4.12.3
requests==2.32.3

#database
psycopg2==2.9.10

